added heavy integration test to base
