+++ base/tests/utils/heavy/src/com/google/idea/blaze/base/InMemoryPreferencesFactory.java
@@ -1,4 +1,4 @@
-package com.google.idea.blaze.clwb.base;
+package com.google.idea.blaze.base;
 
 import java.util.HashMap;
 import java.util.Map;
